+++ lib/src/widget/auto_size_text.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:flutter/material.dart';
+import 'package:material_ui/material_ui.dart';
 
 /// Flutter widget that automatically resizes text to fit perfectly within its
 /// bounds.