+++ lib/src/widget/fitted_text.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_whale/flutter_whale.dart';
-import 'package:flutter/material.dart';
+import 'package:material_ui/material_ui.dart';
 
 class FittedText extends StatelessWidget {
   final String data;